+++ platformio/debug/config/base.py
@@ -145,7 +145,7 @@ class DebugConfigBase:  # pylint: disable=too-many-instance-attributes
         )
 
     def _load_build_data(self):
-        data = load_build_metadata(os.getcwd(), self.env_name, cache=True)
+        data = load_build_metadata(os.getcwd(), self.env_name, cache=True, debug=True)
         if data:
             return data
         raise DebugInvalidOptionsError("Could not load a build configuration")