+++ platformio/telemetry.py
@@ -30,6 +30,7 @@ import requests
 from platformio import __version__, app, exception, util
 from platformio.cli import PlatformioCLI
 from platformio.compat import hashlib_encode_data, string_types
+from platformio.http import HTTPSession
 from platformio.proc import is_ci, is_container
 from platformio.project.helpers import is_platformio_project
 
@@ -146,6 +147,7 @@ class MeasurementProtocol(TelemetryBase):
             "device",
             "org",
             "package",
+            "pkg",
             "platform",
             "project",
             "settings",
@@ -183,7 +185,7 @@ class MeasurementProtocol(TelemetryBase):
     def _ignore_hit(self):
         if not app.get_setting("enable_telemetry"):
             return True
-        if self["ea"] in ("Idedata", "_Idedata"):
+        if self["ea"] in ("Idedata", "__Idedata"):
             return True
         return False
 
@@ -205,7 +207,7 @@ class MPDataPusher:
     def __init__(self):
         self._queue = queue.LifoQueue()
         self._failedque = deque()
-        self._http_session = requests.Session()
+        self._http_session = HTTPSession()
         self._http_offline = False
         self._workers = []
 
@@ -269,7 +271,6 @@ class MPDataPusher:
             r = self._http_session.post(
                 "https://ssl.google-analytics.com/collect",
                 data=data,
-                headers={"User-Agent": app.get_user_agent()},
                 timeout=1,
             )
             r.raise_for_status()